docs: add javadoc to classes in the servicemanager package
