+++ src/main/java/org/example/learningsystem/btp/servicemanager/binding/dto/HanaBindingCredentials.java
@@ -5,16 +5,31 @@ import lombok.Getter;
 import lombok.Setter;
 import org.example.learningsystem.core.db.config.DataSourceProperties;
 
+/**
+ * Represents the credentials for an SAP HANA Database schema binding.
+ */
 @Getter
 @Setter
 public class HanaBindingCredentials implements DataSourceProperties {
 
+    /**
+     * The JDBC URL for connecting to the schema.
+     */
     private String url;
 
+    /**
+     * The username for database authentication.
+     */
     @JsonProperty("user")
     private String username;
 
+    /**
+     * The password for the database authentication.
+     */
     private String password;
 
+    /**
+     * The name of the schema.
+     */
     private String schema;
 }
