+++ src/main/java/org/example/learningsystem/btp/servicemanager/binding/service/ServiceBindingManager.java
@@ -5,7 +5,7 @@ import org.example.learningsystem.btp.servicemanager.binding.dto.CreateServiceBi
 import org.example.learningsystem.btp.servicemanager.binding.dto.ServiceBindingResponseDto;
 import org.example.learningsystem.btp.servicemanager.common.service.BaseServiceManager;
 import org.example.learningsystem.btp.servicemanager.common.util.ServiceManagerRestClient;
-import org.example.learningsystem.btp.servicemanager.common.util.ServiceManagerURIBuilder;
+import org.example.learningsystem.btp.servicemanager.common.builder.ServiceManagerURIBuilder;
 import org.example.learningsystem.multitenancy.context.TenantInfo;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
@@ -19,6 +19,9 @@ import static org.example.learningsystem.btp.servicemanager.common.constant.Serv
 import static org.example.learningsystem.btp.servicemanager.common.constant.ServiceManagerResourceConstants.SUBDOMAIN;
 import static org.example.learningsystem.btp.servicemanager.common.constant.ServiceManagerResourceConstants.TENANT_ID;
 
+/**
+ * Service for managing service bindings in SAP BTP using the Service Manager API.
+ */
 @Service
 @Profile("cloud")
 @RequiredArgsConstructor
@@ -28,6 +31,14 @@ public class ServiceBindingManager {
     private final ServiceManagerRestClient serviceManagerRestClient;
     private final ServiceManagerURIBuilder serviceManagerURIBuilder;
 
+    /**
+     * Creates a new service binding.
+     *
+     * @param name              the name of the binding
+     * @param serviceInstanceId the id of the service instance to bind to
+     * @param tenantInfo        the tenant information
+     * @return the {@link ServiceBindingResponseDto} instance
+     */
     public ServiceBindingResponseDto create(String name, UUID serviceInstanceId, TenantInfo tenantInfo) {
         var uri = serviceManagerURIBuilder.builder(SERVICE_BINDINGS)
                 .async(false)
@@ -40,15 +51,31 @@ public class ServiceBindingManager {
         return serviceManagerRestClient.post(uri, body, ServiceBindingResponseDto.class);
     }
 
+    /**
+     * Retrieves a service binding by its associated tenant id.
+     *
+     * @param tenantId the tenant id to search for
+     * @return a {@link ServiceBindingResponseDto} instance
+     */
     public ServiceBindingResponseDto getByTenantId(String tenantId) {
         return baseServiceManager.getByLabel(TENANT_ID, tenantId, SERVICE_BINDINGS, ServiceBindingResponseDto.class);
     }
 
+    /**
+     * Deletes a service binding by its name.
+     *
+     * @param name the name of the binding.
+     */
     public void deleteByName(String name) {
         var binding = baseServiceManager.getByField(NAME, name, SERVICE_BINDINGS, ServiceBindingResponseDto.class);
         baseServiceManager.deleteById(SERVICE_BINDINGS, binding.id());
     }
 
+    /**
+     * Retrieves all service bindings.
+     *
+     * @return a {@link List} of all service bindings
+     */
     public List<ServiceBindingResponseDto> getAll() {
         var uri = serviceManagerURIBuilder.builder(SERVICE_BINDINGS).build();
         var response = baseServiceManager.getAll(uri, ServiceBindingResponseDto.class);
